chore: cleanup redundant types and remove using fetch
