+++ contract/types.d.ts
@@ -53,14 +53,6 @@ export type AbiEncodedContractCall = {
   data: `0x${string}`;
 };
 
-export const EVM_CHAINS = {
-  Avalanche: 'Avalanche',
-  Base: 'base-sepolia',
-  Ethereum: 'ethereum-sepolia',
-};
-
-export type SupportedDestinationChains = keyof typeof EVM_CHAINS;
-
 export type OfferArgs = {
   destinationAddress: `0x${string}`;
   type: GMPMessageType;
