+++ deploy/test/axelar-gmp.test.ts
@@ -101,7 +101,7 @@ test.before(async (t) => {
   await evalProposal(
     buildProposal('../src/init-contract.js', [
       '--net',
-      'local',
+      'localhost',
       '--peer',
       'axelar:connection-0:channel-0:uaxl',
     ]),
